refactor(tree): update imports and enhance tree component structure
